feat: Input 클래스를 Names와 Repeat로 분리
Input 클래스 삭제
NamesTest 기존 InputTest에서 수행하던 이름 입력 관련 테스트 수행
Names 테스트에 맞춰 기존 Input의 names 관련 필드/메서드 구현
RepeatTest 기존 InputTest에서 수행하던 횟수 입력 관련 테스트 수행
Repeat 테스트에 맞춰 기존 Input의 repeat 관련 필드/메서드 구현
