+++ src/test/java/racinggame/domain/data/OutputTest.java
@@ -1,4 +1,4 @@
-package racinggame.controller;
+package racinggame.domain.data;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
